Some validation added . . .

+++ Controllers/ClientController.cs
@@ -12,22 +12,33 @@ namespace DotNet_6_REST_API_with_mongoDB.Controllers
     [ApiController]
     public class ClientController : ControllerBase
     {
-        // private readonly IConfiguration _configuration;
+
         private readonly MongoDBService _mongoDBService;
-        // public ClientController(IConfiguration configuration)
-        // {
-        //     _configuration = configuration;
-        // }
 
         public ClientController(MongoDBService mongoDBService)
         {
             _mongoDBService = mongoDBService;
         }
 
+        // general guidance is that if you are making database calls, always make that code asynchronous
+
         [HttpGet]
         public async Task<List<Our_Clients>> Get()
         {
             return await _mongoDBService.GetAsync();
+        }
+
+        [HttpGet("{id:length(24)}")]
+
+        public async Task<IActionResult> GetById(string id)
+        {
+
+            var client = await _mongoDBService.GetbyIdAsync(id);
+            if (client == null)
+            {
+                return NotFound();
+            }
+            return Ok(client);
         }
 
         [HttpPost]
@@ -39,22 +50,46 @@ namespace DotNet_6_REST_API_with_mongoDB.Controllers
             client.password = RandomString(8);
             client.date_of_birth = "1998-11-06";
 
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             await _mongoDBService.CreateAsync(client);
 
             //return StatusCode(200);
             return CreatedAtAction(nameof(Get), new { id = client.Id }, client);
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> AddToPlaylist(string id, [FromBody] string client_pass)
+        [HttpPut("{id:length(24)}")]
+        public async Task<IActionResult> AddtoClient(string id, [FromBody] string client_pass)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
+            var client = await _mongoDBService.GetbyIdAsync(id);
+
+            if (client == null)
+            {
+                return NotFound();
+            }
+
             await _mongoDBService.AddToClientsAsync(id, "MasumTheHero");
             return NoContent();
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:length(24)}")]
         public async Task<IActionResult> Delete(string id)
         {
+
+            var client = await _mongoDBService.GetbyIdAsync(id);
+
+            if (client == null)
+            {
+                return NotFound();
+            }
             await _mongoDBService.DeleteAsync(id);
             return NoContent();
         }
